requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.10
matplotlib>=3.8
joblib>=1.3
pyyaml>=6.0
shap>=0.44
